chore: relocate history utxo utils under explorer utils
