+++ src/History/importExportParser.ts
@@ -1,9 +1,9 @@
 import { iHistoryImportExport } from '@/History/types';
 import { parseMemo } from '@/History/history_helpers';
 import { idToChainAlias } from '@/Network/helpers/aliasFromNetworkID';
-import { activeNetwork, xChain } from '@/Network/network';
+import { xChain } from '@/Network/network';
 import { bnToAvaxX } from '@/utils';
-import { getOutputsOfChain, getOutputTotals, getOwnedOutputs } from '@/History/utxo_helpers';
+import { getOutputsOfChain, getOutputTotals, getOwnedOutputs } from '@/Explorer/ortelius/utxoUtils';
 import { findDestinationChain, findSourceChain, OrteliusAvalancheTx } from '@/Explorer';
 
 export function getImportSummary(tx: OrteliusAvalancheTx, addresses: string[]): iHistoryImportExport {
